Unexpected behavior: ONE defaults Settings with "default_settings.txt" when provided with another settings file
Problem description by example:
Let "new_settings.txt" define a new scenario with two groups but only sets a default speed for both groups: "Group.speed = 0.5, 1.5". Since there are no explicit settings for "Group1.speed" or "Group2.speed" once would expect both groups to be set to the default speed when calling './one.sh new_settings.txt'.
This is not the case: Since "default_settings.txt" defines explicitly "Group2.speed = 2.7, 13.9", this (specialized) setting will be applied to group 2 of "new_settings.txt" because the Properties object is defaulted with "default_settings.txt" if it exists.

This is unexpected behavior and should be removed. Note that if "default_settings.txt" is deleted or directly changed, this is not a problem. There are no warning messages whether "default_settings.txt" is used or not.