Introduce transfer and delivery queues to speed up simulation
ActiveRouter#update() iterates over all connections and messages and tries to transfer them, which causes excessive overhead (#exchangeDeliverableMessages() and #tryAllMessagesToAllConnections()). The longest runtime is observed when all routers are in sync as all messages are tried on all connections every update interval (each attempt failes as the other router already has the message)

This patch introduces transfer and delivery queues which are maintained for active connections. This way, ActiveRouter will resume the message--connection iteration on update() instead of starting over.

First tests attest a speed up of about 50x.